refactor: StationList -> Stations 일급 객체로 변경
