+++ src/main/java/nextstep/subway/path/PathFinder.java
@@ -1,6 +1,6 @@
 package nextstep.subway.path;
 
-import nextstep.subway.line.section.Lines;
+import nextstep.subway.line.Lines;
 import nextstep.subway.station.Station;
 
 public interface PathFinder {
